requirements.txt
appdirs==1.4.4
asgiref==3.2.7
certifi==2020.12.5
distlib==0.3.0
Django==3.0
django-classy-tags==1.0.0
django-filter==2.4.0
django-js-asset==1.2.2
filelock==3.0.12
gunicorn==20.0.4
packaging==20.3
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyparsing==3.0.0a1
pytz==2019.3
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
static3==0.7.0
virtualenv==20.0.21
virtualenv-clone==0.5.4
webencodings==0.5.1
whitenoise==5.2.0
zipp==3.1.0
